Перенесла функцию в data.js
